Main algorithm finished
This is the final form of the generalized A* pathfinding algorithm, with comments. Specifics that still need to be added to fit our use case include:

Distance function that distinguishes stairs/z-axis. Relating abstract Nodes to searchable locations on campus (BF 115).

Data structure for storing nodes that allows us to efficiently find the closest node, given an (x,y,z) position