Began API Function Dev
Created a function that takes two strings for a user current and destination locations and makes a POST request to API endpoint with serialized data. Function awaits response from server with path and returns it once it receives it.
Also added comments to some other files while reading through them