Create Conversion.cs
I don't know how to get C# code to run in the actual repository. So I wrote it in a personal repository, and now I'm just copying the file over. It should be done/working. The Coordinate object will need changed to accept Nodes from the database, but it should be relatively minor stuff.