--------> DISGRAD LAST MESSAGE
Added a second search bar for room number
alongside the existing one that is used for
building. Also reformatted the search bar.
Added functionality so users can click the search
list and it fills search bar with clicked item.
Hides the search list if a user clicked an item
from the list, redisplays it if the user changes
their mind and starts backspacing.